requirements.txt
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
confusable-homoglyphs==3.3.1
Django==6.0.1
django-filter==25.2
django-registration==5.2.1
djangorestframework==3.16.1
dnspython==2.8.0
drf-yasg==1.21.14
geojson==3.2.0
h11==0.16.0
idna==3.11
inflection==0.5.1
Markdown==3.10.1
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
panda==0.3.1
pandas==2.3.3
pillow==12.1.0
plyer==2.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pygame==2.6.1
pymongo==4.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
selenium==4.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.5
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
